# dist/courier-api-swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: CourierExpress OpenAPI
  description: ''
  contact: {}
host: localhost:8080
basePath: /api
securityDefinitions: {}
schemes:
- "https"
- "http"
consumes:
- application/json
produces:
- application/json
paths:
...
  /user/address:
    get:
      description: ''
      summary: Get All Addresses
      tags:
      - Address Routes
      operationId: Address
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GetAddressResponse'
          headers: {}
        '404':
          description: Page not found
        '400':
          description: No token found
      security: []
    post:
      description: ''
      summary: Post User Address
      tags:
      - Address Routes
      operationId: PostAddress
      deprecated: false
      produces:
      - application/json
      consumes:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/AddressRequest'
      - name: Content-Type
        in: header
        required: false
        enum:
        - application/json
        type: string
        description: ''
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/CreateAddressResponse'
          headers: {}
        '404':
          description: Page not found
        '400':
          description: No token found
      security: []
...
  /user/user-promo:
    get:
      description: ''
      summary: User Promos
      tags:
      - User Promo Routes
      operationId: UserPromos
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GetAllUserPromosResponse'
          headers: {}
        '404':
          description: Page not found
        '400':
          description: No token found
      security: []
    post:
      description: ''
      summary: User Promos
      tags:
      - User Promo Routes
      operationId: PostUserPromos
      deprecated: false
      produces:
      - application/json
      consumes:
      - application/json
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/UserPromosRequest'
      - name: Content-Type
        in: header
        required: false
        enum:
        - application/json
        type: string
        description: ''
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PostUserPromoResponse'
          headers: {}
        '404':
          description: Page not found
        '400':
          description: No token found
      security: []
...
  /admin/promo:
    get:
      description: ''
      summary: All Promos
      tags:
      - Admin Manage Promos
      operationId: AllPromos
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GetAllPromosResponse'
          headers: {}
        '404':
          description: Page not found
      security: []
  /admin/user/referrals:
    get:
      description: ''
      summary: All Users
      tags:
      - Admin Manage Referrals
      operationId: AllUsers
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: Authorization
        in: header
        required: true
        type: string
        description: ''
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GetReferralsResponse'
          headers: {}
        '404':
          description: Page not found
        '400':
          description: No token found
      security: []
definitions:
...
  AddressRequest:
    title: AddressRequest
    example:
      full_address: Jl. Sopo Del
      recipient_name: Rizky Ashari
      recipient_phone_number: 6285722326522
    type: object
    properties:
      full_address:
        type: string
      recipient_name:
        type: string
      recipient_phone_number:
        type: integer
        format: int64
    required:
    - full_address
    - recipient_name
    - recipient_phone_number
...
  UserPromosRequest:
    title: UserPromosRequest
    example:
      user_id: 2
      promo_id: 2
    type: object
    properties:
      user_id:
        type: integer
        format: int32
      promo_id:
        type: integer
        format: int32
    required:
    - user_id
    - promo_id
...
  CreateAddressResponse:
    title: CreateAddressResponse
    example:
      id: 11
      full_address: Jl. Sopo Del
      recipient_name: Rizky Ashari
      recipient_phone_number: 6285722326522
      CreatedAt: 2022-12-02T18:50:53.795335+07:00
      UpdatedAt: 2022-12-02T18:50:53.795335+07:00
    type: object
    properties:
      id:
        type: integer
        format: int64
      full_address:
        type: string
      recipient_name:
        type: string
      recipient_phone_number:
        type: integer
        format: int64
      CreatedAt:
        type: string
      UpdatedAt:
        type: string
...
  GetAddressResponse:
    title: GetAddressResponse
    example:
      status: true
      message: OK
      errors: null
      data: [
       {
        id: 11,
        user: {
          id: 4,
          name: Doni Ahmad,
          email: doniahmad@email.com,
          phone_number: 6285722320622,
          role: User,
          referral_code: DONI3061,
          balance: 0,
          photos: https://res.cloudinary.com/dmlzx9yxe/image/upload/v1669964733/wazdcvhu1lzoqp6v6lqo.jpg,
          },
        full_address: Jl. Sopo Del,
        recipient_name: Rizky Ashari,
        recipient_phone_number: 6285722326522,
        CreatedAt: 2022-12-02T18:50:53.795335+07:00,
        UpdatedAt: 2022-12-02T18:50:53.795335+07:00
       },
       {
        id: 14,
        user: {
          id: 4,
          name: Doni Ahmad,
          email: doniahmad@email.com,
          phone_number: 6285722320622,
          role: User,
          referral_code: DONI3061,
          balance: 0,
          photos: https://res.cloudinary.com/dmlzx9yxe/image/upload/v1669964733/wazdcvhu1lzoqp6v6lqo.jpg,
          },
        full_address: Jl. Rawa Simprug,
        recipient_name: Andi Mallarangeng,
        recipient_phone_number: 6285788792345,
        CreatedAt: 2022-12-03T11:48:18.490002+07:00,
        UpdatedAt: 2022-12-03T11:48:18.490002+07:00
       }
      ]
    type: object
    properties:
      status:
        type: boolean
      message:
        type: string
      errors:
        type: object
      data:
        type: array
...
  PostUserPromoResponse:
    title: PostUserPromoResponse
    example:
      status: true
      message: OK
      errors: null
      data: {
        id: 218,
        user: {
          id: 3,
          name: "Dedy Irawan",
          email: "dedy@gmail.com",
          phone_number: 6285722329090,
          role: "User",
          referral_code: "DEDY8190",
          balance: 10360000,
          photos: "https://res.cloudinary.com/dmlzx9yxe/image/upload/v1669964809/rjbj2ccw3acf2ml0ouey.png"
        },
        promo: {
          id: 2,
          name: "60% discount vouchers",
          min_fee: 20000,
          max_discount: 15000,
          discount: 60,
          quota: 40
        },
        status: 0
          }
    type: object
    properties:
      status:
        type: boolean
      message:
        type: string
      errors:
        type: object
      data:
        type: object
  GetAllUserPromosResponse:
    title: GetAllUserPromosResponse
    example:
      status: true
      message: "OK"
      errors: null
      data: [
        {
        id: 218,
        user: {
          id: 3,
          name: "Dedy Irawan",
          email: "dedy@gmail.com",
          phone_number: 6285722329090,
          role: "User",
          referral_code: "DEDY8190",
          balance: 10360000,
          photos: "https://res.cloudinary.com/dmlzx9yxe/image/upload/v1669964809/rjbj2ccw3acf2ml0ouey.png"
        },
        promo: {
          id: 2,
          name: "60% discount vouchers",
          min_fee: 20000,
          max_discount: 15000,
          discount: 60,
          quota: 40
        },
        status: 0
       },
        {
          id: 219,
          user: {
            id: 3,
            name: "Dedy Irawan",
            email: "dedy@gmail.com",
            phone_number: 6285722329090,
            role: "User",
            referral_code: "DEDY8190",
            balance: 10360000,
            photos: "https://res.cloudinary.com/dmlzx9yxe/image/upload/v1669964809/rjbj2ccw3acf2ml0ouey.png"
        },
        promo: {
          id: 3,
          name: "80% discount voucher uhuyy",
          min_fee: 20000,
          max_discount: 15000,
          discount: 80,
          quota: 120
        },
        status: 0
        },
       ]
    type: object
    properties:
      status:
        type: boolean
      message:
        type: string
      errors:
        type: object
      data:
        type: array
...
  GetAllPromosResponse:
    title: GetAllPromosResponse
    example:
      status: true
      message: Deleted
      errors: null
      data: [
        {
          id: 1,
          name: "40% discount vouchers",
          min_fee: 20000,
          max_discount: 15000,
          discount: 40,
          quota: 100
        },
        {
          id: 2,
          name: "60% discount vouchers",
          min_fee: 20000,
          max_discount: 15000,
          discount: 60,
          quota: 40
        },
        {
          id: 3,
          name: "80% discount voucher uhuyy",
          min_fee: 20000,
          max_discount: 15000,
          discount: 80,
          quota: 120
        }
      ]
    type: object
    properties:
      status:
        type: boolean
      message:
        type: string
      errors:
        type: object
      data:
        type: object
  GetReferralsResponse:
    title: GetReferralsResponse
    example:
      status: true
      message: Deleted
      errors: null
      data: [
        "DEDY8190",
        "DONI3798",
        "DONI3061",
        "RIZK0967",
        "AHMA8013",
        "DONI9250",
        "ANDI0659"
        ]
    type: object
    properties:
      status:
        type: boolean
      message:
        type: string
      errors:
        type: object
      data:
        type: array
security: []
...